requirements.txt
cuid
gitpython
giturlparse
questionary
pydantic-settings
randomname
sentry-sdk
tinydb
typer[all]
uWSGI
uwsgiconf
uwsgi-tasks @ git+https://github.com/EloquentBits/uwsgi_tasks@master
#zmq
